% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cetesb_station_ids.R
\docType{data}
\name{cetesb_station_ids}
\alias{cetesb_station_ids}
\title{Ids for the CETESB stations.}
\format{A data frame with 62 rows and 3 variables:
\describe{
\item{id}{Id of the station.}
\item{stationname}{Name of the station.}
\item{initial_date}{Date when the station started to operate,
in format dd/mm/aaaa.}
}}
\source{
\url{http://qualar.cetesb.sp.gov.br/}
}
\usage{
cetesb_station_ids
}
\description{
A dataset containing the ids for every CETESB station.
}
\keyword{datasets}
