% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_cetesb.R
\name{scraper_cetesb}
\alias{scraper_cetesb}
\title{Scraper CETESB}
\usage{
scraper_cetesb(station, parameter, start, end, login, password, type = "P",
  invalidData = "on", network = "A", cleanData = TRUE)
}
\arguments{
\item{station}{A numeric value indicating the station id from
where you wish to get the data.
See \link[koffing:cetesb_station_ids]{koffing::cetesb_station_ids}.}

\item{parameter}{A numeric value indicating the CETESB parameter
id. See \link[koffing:cetesb_station_ids]{koffing::cetesb_station_ids}.}

\item{start}{A string in the format "dd/mm/aaaa" representing
the initial day for the data selection.}

\item{end}{A string in the format "dd/mm/aaaa" representing  the
final day for the data selection.}

\item{login}{A string with your login on Qualar system.}

\item{password}{A string with your passoword on Qualar system.}

\item{type}{Type of data: "P" for hourly mean or "M" for moving
average.}

\item{invalidData}{If TRUE, the system will return rows with
invalid data.}

\item{network}{Network type: "A" for automatic or "M" for manual.}

\item{cleanData}{Should the data be cleaned before returned? The
default is TRUE. Set FALSE to return the data as extracted from Qualar.}
}
\value{
A tibble with the data returned by the Qualar system.
}
\description{
Scraper CETESB
}
\examples{
\dontrun{

# Ozone for 'Dom Pedro II' station from 01/01/2018 to 31/01/2018.

scraper_cetesb(station = 72, parameter = 63,
               start = "01/01/2018", end = "31/01/2018",
               login = "my_login", password = "my_password")
}
}
